refact: rename folders & files in cebab-case
