--- src/common/hooks/useFetch/index.js
@@ -1,3 +1,0 @@
-import { useFetch } from './useFetch';
-
-export { useFetch };
