--- src/common/hooks/useFetchReducer/index.js
@@ -1,3 +1,0 @@
-import { useFetchReducer } from './useFetchReducer';
-
-export { useFetchReducer };
