--- src/components/CheckoutOrderDetails/index.js
@@ -1,2 +1,0 @@
-import { CheckoutOrderDetails } from './CheckoutOrderDetails';
-export { CheckoutOrderDetails };
