+++ src/components/Dropdown/index.js
@@ -1,2 +1,3 @@
 import { Dropdown } from './dropdown';
+
 export { Dropdown };
