--- src/components/ErrorModalDetails/index.js
@@ -1,2 +1,0 @@
-import { ErrorModalDetails } from './ErrorModalDetails';
-export { ErrorModalDetails };
