--- src/components/LoaderShape/index.js
@@ -1,2 +1,0 @@
-import { LoaderShape } from './LoaderShape';
-export { LoaderShape };
