+++ src/components/burger-constructor-total/burger-constructor-total.js
@@ -6,8 +6,8 @@ import { Price } from '../price';
 import { Modal } from '../modal';
 import { OrderDetails } from '../order-details';
 import { useIntl } from 'react-intl';
-import { CheckoutOrderDetails } from '../CheckoutOrderDetails';
-import { ErrorModalDetails } from '../ErrorModalDetails';
+import { CheckoutOrderDetails } from '../checkout-order-details';
+import { ErrorModalDetails } from '../error-modal-details';
 import { FETCH_STATUS } from '../../utils/constants';
 import { AnimatePresence } from 'framer-motion';
 import { useDispatch, useSelector } from 'react-redux';
