+++ src/components/card/price.js
@@ -1,5 +1,5 @@
 import { Price as PriceBasic } from '../price';
-import { useCardContext } from './hook/useCardContext';
+import { useCardContext } from './hook/use-card-context';
 
 export const Price = () => {
   const product = useCardContext();
