+++ src/components/checkout-order-details/index.js
@@ -1,0 +1,2 @@
+import { CheckoutOrderDetails } from './checkout-order-details';
+export { CheckoutOrderDetails };
