+++ src/components/error-modal-details/error-modal-details.module.scss
