+++ src/components/error-modal-details/index.js
@@ -1,0 +1,2 @@
+import { ErrorModalDetails } from './error-modal-details';
+export { ErrorModalDetails };
