+++ src/components/loader-shape/index.js
@@ -1,0 +1,2 @@
+import { LoaderShape } from './loader-shape';
+export { LoaderShape };
